fix(metrics): cast class_id to float32 in detections_to_tensor
Concatenating int32 class_id with float32 box coordinates causes numpy
to upcast the whole result tensor to float64, breaking the declared
float32 return type. Explicit cast to float32 before concatenation
preserves the contract.

---

+++ src/supervision/metrics/detection.py
@@ -127,7 +127,10 @@ def detections_to_tensor(
     else:
         box_data = detections.xyxy
 
-    arrays_to_concat = [box_data, np.expand_dims(detections.class_id, 1)]
+    arrays_to_concat = [
+        box_data,
+        np.expand_dims(detections.class_id.astype(np.float32), 1),
+    ]
 
     if with_confidence:
         if detections.confidence is None:
